refactor(emulator): move Modbus protocol helpers to utils folder
Moved modbus-protocol-helpers.ts and its tests from transports/ to utils/
folder for better organization. These are utility functions that are not
transport-specific and belong in the utils directory.

Changes:
- Moved modbus-protocol-helpers.ts to utils/
- Moved modbus-protocol-helpers.test.ts to utils/
- Updated imports in tcp.ts and rtu.ts to point to new location

All 246 tests pass, verifying no behavioral changes.
