+++ packages/emulator/src/utils/modbus-protocol-helpers.test.ts
