+++ packages/emulator/src/utils/modbus-protocol-helpers.ts
